Consider all transactions in both new and existing blocks during indexing
This is a shift away from being driven by account history instead of individual blocks.
Going through all the blocks also makes it possible to process a wider range of transaction types.

In addition to PoolSwap, the transaction types AddPoolLiquidity and RemovePoolLiquidity are now also processed but not yet stored in the database.